+++ graphannis/src/annis/db/corpusstorage.rs
@@ -52,6 +52,7 @@ use sys_info;
 
 use anyhow::{Context, Error};
 use aql::model::AnnotationComponentType;
+use brotli::CompressorWriter;
 use db::AnnotationStorage;
 
 #[cfg(test)]
@@ -246,6 +247,8 @@ pub enum ImportFormat {
     /// [GraphML](http://graphml.graphdrawing.org/) based export-format, suitable to be imported from other graph databases.
     /// This format follows the extensions/conventions of the Neo4j [GraphML module](https://neo4j.com/docs/labs/apoc/current/import/graphml/).
     GraphML,
+    /// Like `GraphML`, but compressed with the [Brotli](https://github.com/google/brotli) compression format.
+    GraphMLCompressed,
 }
 
 /// An enum of all supported output formats of graphANNIS.
@@ -255,6 +258,8 @@ pub enum ExportFormat {
     /// [GraphML](http://graphml.graphdrawing.org/) based export-format, suitable to be imported into other graph databases.
     /// This format follows the extensions/conventions of the Neo4j [GraphML module](https://neo4j.com/docs/labs/apoc/current/import/graphml/).
     GraphML,
+    /// Like `GraphML`, but compressed with the [Brotli](https://github.com/google/brotli) compression format.
+    GraphMLCompressed,
 }
 
 /// Different strategies how it is decided when corpora need to be removed from the cache.
@@ -731,7 +736,7 @@ impl CorpusStorage {
         let (orig_name, mut graph) = match format {
             ImportFormat::RelANNIS => relannis::load(path, disk_based, |status| {
                 info!("{}", status);
-                // loading the file from relANNIS consumes memory, update the corpus cache regulary to allow it to adapat
+                // loading the file from relANNIS consumes memory, update the corpus cache regularly to allow it to adapt
                 self.check_cache_size_and_remove(vec![]);
             })?,
             ImportFormat::GraphML => {
@@ -743,6 +748,30 @@ impl CorpusStorage {
                 let input_file = File::open(path)?;
                 let g = graphannis_core::graph::serialization::graphml::import(
                     input_file,
+                    disk_based,
+                    |status| {
+                        info!("{}", status);
+                        // loading the file from relANNIS consumes memory, update the corpus cache regulary to allow it to adapat
+                        self.check_cache_size_and_remove(vec![]);
+                    },
+                )?;
+                (orig_corpus_name, g)
+            }
+            ImportFormat::GraphMLCompressed => {
+                let orig_corpus_name = if let Some(file_name) = path.file_stem() {
+                    let mut file_name = file_name.to_string_lossy().to_string();
+                    // The file might end with ".br" as file ending instead of ".graphml"
+                    if file_name.ends_with(".graphml") {
+                        file_name.truncate(file_name.len() - ".graphml".len());
+                    }
+                    file_name
+                } else {
+                    "UnknownCorpus".to_string()
+                };
+                let input_file = File::open(path)?;
+                let decompressed_input = brotli::Decompressor::new(input_file, 4096);
+                let g = graphannis_core::graph::serialization::graphml::import(
+                    decompressed_input,
                     disk_based,
                     |status| {
                         info!("{}", status);
@@ -834,6 +863,13 @@ impl CorpusStorage {
                         info!("{}", status);
                     },
                 )?;
+            }
+            ExportFormat::GraphMLCompressed => {
+                let output_file = File::create(path)?;
+                let writer = CompressorWriter::new(output_file, 4096, 9, 22);
+                graphannis_core::graph::serialization::graphml::export(graph, writer, |status| {
+                    info!("{}", status);
+                })?;
             }
         }